django-restframe-work/company_api/requirements.txt
asgiref==3.8.1
django==5.0.6
djangorestframework==3.15.2
sqlparse==0.5.0
